+++ changelog.md
@@ -4,6 +4,7 @@
 * Fix - Recomplied Minified JS
 * Dev - Added in a filter to disable the LSX Banners if the page is using Blocks
 * Fix - Making sure the Theme is pre 5.0 compatible
+* Fix - Fixing the breadcrumb for The Events Calendar Community Event when using Yoast SEO Breadcrumbs 
 
 ### 2.2.1
 * Fix - Fixed the white screen error when not using the Gutenburg editor.